% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{coerce-spatPomp-data.frame}
\alias{coerce,spatPomp,data.frame-method}
\alias{as.data.frame.spatPomp}
\title{Coerce to data frame}
\usage{
\method{as.data.frame}{spatPomp}(x, ...)
}
\arguments{
\item{x}{any \R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
\pkg{spatPomp} objects can be recast as data frames.
}
\details{
When \code{object} is a simple \sQuote{spatPomp} object,
\code{as(object,"data.frame")} or \code{as.data.frame(object)} results in a
data frame with the times, units, observables, states (if known), and
interpolated covariates (if any).
}
