% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm.R
\name{bm}
\alias{bm}
\title{Brownian motion spatpomp generator}
\usage{
bm(U = 5, N = 100, delta.t = 0.1)
}
\arguments{
\item{U}{A length-one numeric signifying dimension of the process.}

\item{N}{A length-one numeric signifying the number of time steps to evolve the process.}
}
\value{
A spatpomp object with the specified dimension and time steps.
}
\description{
Generate a spatpomp object representing a \code{U}-dimensional  
Brownian motion with spatial correlation decaying geometrically with
distance around a circle. The model is defined in continuous time
though in this case an Euler approximation is exact at the evaluation
times.
}
\examples{
bm(U=4, N=20)
}
