% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_data_UK.R
\name{city_data_UK}
\alias{city_data_UK}
\title{City data in the United Kingdom}
\description{
Population and birth information about cities in England and Wales during the measles pre-vaccine era.
}
\details{
Data includes births and population at bi-weekly observations from 40 cities and towns.
}
\references{
\dalziel2016
}
\seealso{
Other datasets: 
\code{\link{measlesUK}}
}
\concept{datasets}
