% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc.R
\name{conc}
\alias{conc}
\alias{conc,SpatPomp-method}
\alias{conc,Bpfilter-method}
\alias{conc,Ibpf-method}
\title{Concatenate}
\usage{
\S4method{conc}{SpatPomp}(...)

\S4method{conc}{Bpfilter}(...)

\S4method{conc}{Ibpf}(...)
}
\description{
Internal methods to concatenate objects into useful listie.
}
\details{
Not exported.
}
\keyword{internal}
