% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_extended.R
\name{ebm}
\alias{ebm}
\title{Brownian motion spatPomp generator}
\usage{
ebm(U = 5, N = 100, rw.sd = 0.02, cooling.fraction.50 = 0.5, delta.t = 0.1)
}
\arguments{
\item{U}{A length-one numeric signifying dimension of the process.}

\item{N}{A length-one numeric signifying the number of observation time steps to evolve the process.}
}
\value{
A spatPomp object with the specified dimension and time steps.
}
\description{
Generate a spatPomp object representing a \code{U}-dimensional
Brownian motion with spatial correlation decaying geometrically with
distance around a circle. The model is defined in continuous time
though in this case an Euler approximation is exact at the evaluation
times.
}
\examples{
b <- bm(U=4, N=20)
# See all the model specifications of the object
spy(b)
}
