% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girf.R
\docType{methods}
\name{girf}
\alias{girf}
\alias{girf-spatpomp}
\alias{girf,spatpomp-method}
\alias{girf-girfd_spatpomp}
\alias{girf,girfd_spatpomp-method}
\title{Guided intermediate resampling filter (GIRF)}
\usage{
\S4method{girf}{spatpomp}(data, Np, Ninter, lookahead, Nguide, h,
  theta.to.v, v.to.theta, tol = 1e-17, max.fail = Inf,
  save.states = FALSE, ..., verbose = getOption("verbose", FALSE))

\S4method{girf}{girfd_spatpomp}(data, Np, tol, Ninter, Nguide, lookahead,
  h, theta.to.v, v.to.theta, ..., verbose = getOption("verbose", FALSE))
}
\arguments{
\item{data}{either a data frame holding the spatiotemporal data,
or an object of class \sQuote{spatPomp},
i.e., the output of another \pkg{spatPomp} calculation.}

\item{Np}{The number of particles used.}

\item{Ninter}{the number of intermediate resampling timepoints.}

\item{lookahead}{The number of future observations included in the guide function.}

\item{Nguide}{The number of simulations used to estimate state process uncertainty for each particle.}

\item{h}{A function with template \code{function(state.vec, param.vec)} which returns an approximation to the expected observed value given a latent state and parameter vector.}

\item{theta.to.v}{A function with template \code{function(state.vec, param.vec)} which returns an approximation to the variance of an observation given a latent state and parameter vector.}

\item{v.to.theta}{A function with template \code{function(var, state.vec, param.vec)} which returns a parameter vector having observation noise consistent with variance \code{var} at latent state \code{state.vec} with other paramters given by \code{param.vec}.}

\item{...}{
    Any additional arguments given to \code{pomp} will be made available to each of the basic components.
    To prevent errors due to misspellings, a warning is issued if any such arguments are detected.
  }

\item{object}{A \code{spatpomp} object.}

\item{params}{A parameter set for the spatiotemporal POMP.}
}
\value{
Upon successful completion, \code{girf} returns an object of class
\sQuote{girfd.spatpomp}.
}
\description{
An implementation of the algorithm of Park and Ionides (2019),
following the pseudocode in Asfaw, Ionides and King (2019).
}
\section{Methods}{

The following methods are available for such an object:
\describe{
\item{\code{\link{loglik}}}{ yields a biased estimate of the log-likelihood of the data under the model. }
}
}

\references{
\park2019

\asfaw2019
}
\seealso{
Other particle filter methods: \code{\link{asif}},
  \code{\link{hippie}}, \code{\link{igirf}}

Other \pkg{spatpomp} filtering methods: \code{\link{asif}},
  \code{\link{igirf}}
}
\concept{\pkg{spatpomp} filtering methods}
\concept{particle filter methods}
