% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he10coordinates.R
\name{he10coordinates}
\alias{he10coordinates}
\title{City data in the United Kingdom}
\value{
a \sQuote{data.frame} of longitude and lattitude for each town.
}
\description{
Longitude and lattitude for the 20 towns in England and Wales studied by
He et al (2010).
}
\references{
\he2010
}
\seealso{
Other datasets he10: 
\code{\link{he10demography}},
\code{\link{he10measles}},
\code{\link{he10mle}}
}
\concept{datasets he10}
