% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he10demography.R
\name{he10demography}
\alias{he10demography}
\title{Demographic data for 20 towns in the United Kingdom}
\value{
a \sQuote{data.frame} of with variables town, year, pop and births.
}
\description{
Population and birth information for some towns in England and Wales
during the measles pre-vaccine era.
}
\details{
Data are annual statistics for the 20 towns analyzed by He et al (2010).
}
\references{
\he2010
}
\seealso{
Other datasets he10: 
\code{\link{he10coordinates}},
\code{\link{he10measles}},
\code{\link{he10mle}}
}
\concept{datasets he10}
