% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he10measles.R
\name{he10measles}
\alias{he10measles}
\title{Measles in the United Kingdom}
\value{
a \sQuote{data.frame} of reported measles cases for 20 towns,
analyzed by He et al (2010).
}
\description{
Measles case data from various cities and towns in England and Wales during the pre-vaccine era.
}
\details{
Data are weekly case counts for the 20 towns analyzed by He et al (2010).
}
\references{
\he2010
}
\seealso{
Other datasets he10: 
\code{\link{he10coordinates}},
\code{\link{he10demography}},
\code{\link{he10mle}}
}
\concept{datasets he10}
