% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he10mle.R
\name{he10mle}
\alias{he10mle}
\title{Measles in the United Kingdom: MLE from He et al (2010)}
\value{
a \sQuote{data.frame} containing the estimated parameters.
}
\description{
Maximum likelihood estimate for fitting a susceptible-exposed-infected-
recovered model to the measles case report data analyzed by He et al (2010).
The values are similar, but not identical, to those reported by He et al.
}
\references{
\he2010
}
\seealso{
Other datasets he10: 
\code{\link{he10coordinates}},
\code{\link{he10demography}},
\code{\link{he10measles}}
}
\concept{datasets he10}
