% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ienkf.R
\docType{class}
\name{ienkf-spatPomp}
\alias{ienkf-spatPomp}
\alias{ienkf,spatPomp-method}
\title{Iterated ensemble Kalman filter (IEnKF)}
\arguments{
\item{Nenkf}{number of iterations of perturbed ENKF.}
}
\value{
Upon successful completion, \code{ienkf} returns an object of class
\sQuote{ienkfd_spatPomp}.
}
\description{
A parameter estimation algorithm combining that uses EnKF to
perform the filtering step in the iterated filtering context of
IF2.
}
\section{Methods}{

The following methods are available for such an object:
\describe{
\item{\code{\link{coef}}}{ gives the Monte Carlo estimate of the maximum likelihood. }
}
}

\examples{
# Create a simulation of a GBM using a default parameter set
gbm10 <- gbm(U = 10, N = 30)

# Set the initial estimates for the unknown parameters
 start_params <- c("rho" = 0.7, "sigma"= 0.5, "tau"=0.5, "X1_0"=1, "X2_0"=1,
"X3_0"=1, "X4_0"=1, "X5_0"=1, "X6_0"=1, "X7_0"=1, "X8_0"=1, "X9_0"=1, "X10_0"=1)

# Run IENKF with the specified parameters
ienkf_out <- ienkf(gbm10,
                   Nenkf = 10,
                   rw.sd = rw.sd(
  rho=0.02, sigma=0.02, tau=0.02,
  X1_0=0.0, X2_0=0.0, X3_0=0.0, X4_0=0.0, X5_0=0.0,
  X6_0=0.0, X7_0=0.0, X8_0=0.0, X9_0=0.0, X10_0=0.0
                   ),
                   cooling.type = "geometric",
                   cooling.fraction.50 = 0.5,
                   Np=1000)

# Get the parameter estimates from the output
coef(ienkf_out)

}
\references{
Evensen, G. (1994) Sequential data assimilation with a
nonlinear quasi-geostrophic model using Monte Carlo methods to forecast
error statistics Journal of Geophysical Research: Oceans 99:10143--10162

Evensen, G. (2009) Data assimilation: the ensemble Kalman filter
Springer-Verlag.

Anderson, J. L. (2001) An Ensemble Adjustment Kalman Filter for Data
Assimilation Monthly Weather Review 129:2884--2903
}
\seealso{
Other particle filter methods: 
\code{\link{abfir}()},
\code{\link{abf}()},
\code{\link{bpfilter-spatPomp}},
\code{\link{girf}()},
\code{\link{igirf}()}

Other \pkg{spatPomp} parameter estimation methods: 
\code{\link{enkf}()}
}
\concept{\pkg{spatPomp} parameter estimation methods}
\concept{particle filter methods}
