% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iif.R
\name{iif}
\alias{iif}
\title{Adapted Simulation Iterated Island Filter (ASIIF)}
\arguments{
\item{object}{A \code{spatpomp} object.}

\item{params}{A parameter set for the spatiotemporal POMP.}

\item{Np}{The number of particles used within each island for the adapted simulations.}

\item{nbhd}{A function which when given a point in space-time \code{(d,n)} will output a 0-1 matrix of points in space-time considered to be
in the neighborhood of \code{(d,n)}.}

\item{islands}{The number of islands for the adapted simulations.}
}
\value{
Upon successful completion, \code{iif} returns an object of class
\sQuote{iifd.pomp}.
}
\description{
An algorithm for estimating the filter distribution of a spatiotemporal partially-observed Markov process (SpatPOMP for short).
Running \code{iif} causes the algorithm to run independent island jobs which each yield an imperfect adapted simulation. Simulating from the "adapted filter"
distribution runs into a curse of dimensionality (COD) problem, which is mitigated by keeping particles in each island close to each other through resampling down
to one particle per island at each observation time point.
The adapted simulations are then weighted in a way that tries to avert COD by making a weak coupling assumption to get an approximate filter distribution.
As a by-product, we also get a biased estimate of the likelihood of the data.
}
\section{Methods}{

The following methods are available for such an object:
\describe{
\item{\code{\link{loglik}}}{ yields a biased estimate of the log-likelihood of the data under the model. }
}
}

\seealso{
Other particle filter methods: \code{\link{hippie}}
}
\concept{\pkg{spatpomp} filtering methods}
\concept{particle filter methods}
