% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measlesUK.R
\name{measlesUK}
\alias{measlesUK}
\title{Measles in the United Kingdom}
\description{
Measles case data from various cities and towns in England and Wales during the pre-vaccine era.
}
\details{
Data includes bi-weekly case counts as well as births and population from 40 cities and towns.
}
\references{
\dalziel2016
}
\seealso{
Other datasets: 
\code{\link{city_data_UK}}
}
\concept{datasets}
