% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ou.R
\name{ou}
\alias{ou}
\title{Linear-Gaussian process generator}
\usage{
ou(D = 5, N = 100)
}
\arguments{
\item{D}{Numeric signifying dimension of the process.}

\item{N}{Numeric signifying the number of time steps to evolve the process.}
}
\value{
A spatpomp object with the specified dimension and time steps.
}
\description{
Generate a spatpomp object representing a \code{D}-dimensional Ornstein-Uhlenbeck process with \code{N} time steps.
}
\examples{
ou(5, 100)
}
