% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot-igirf}
\alias{plot-igirf}
\alias{plot}
\alias{igirf-method}
\alias{igirfd_spatPomp-method}
\alias{plot-spatPomp}
\alias{plot,}
\title{Plotting output of \code{igirf()}}
\usage{
\S4method{plot}{igirfd_spatPomp}(x, params = names(coef(x)), ncol = 3)

\S4method{plot}{spatPomp}(x, log = F)
}
\arguments{
\item{x}{a \code{spatPomp} object}

\item{params}{the names of the parameters which should be in the plot}

\item{ncol}{the number of columns in the grid plot}

\item{log}{should the data be log-transformed before plotting?
This helps in contexts where there are spikes that could take away
attention from the dynamics illustrated by the rest of the data.}
}
\description{
Diagnostic plot for \code{igirf()}

Visualize \code{spatPomp} data
}
