% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print-spatPomp}
\alias{print,spatPomp-method}
\title{Print methods}
\usage{
\S4method{print}{spatPomp}(x)
}
\arguments{
\item{x}{a \code{spatPomp} object}
}
\value{
An object of class \sQuote{spatPomp} is returned *invisibly*. The user is notified on the console only the class of the object. Use \code{spy()} to see model components.
}
\description{
Prints its argument.
}
