% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enkf.R
\docType{methods}
\name{senkf}
\alias{senkf}
\alias{senkf,ANY-method}
\alias{senkf,missing-method}
\alias{senkf-spatpomp}
\alias{senkf,spatpomp-method}
\title{Ensemble Kalman filters}
\usage{
\S4method{senkf}{spatpomp}(data, Np, h, R, ...,
  verbose = getOption("verbose", FALSE))
}
\arguments{
\item{Np}{the number of particles to use.}

\item{h}{function returning the expected value of the observation given the
state.}

\item{R}{matrix; variance of the measurement noise.}

\item{C}{matrix converting state vector into expected value of the
observation.}
}
\value{
An object of class \sQuote{kalmand_spatpomp}.
}
\description{
The ensemble Kalman filter
}
\references{
Evensen, G. (1994) Sequential data assimilation with a
nonlinear quasi-geostrophic model using Monte Carlo methods to forecast
error statistics Journal of Geophysical Research: Oceans 99:10143--10162

Evensen, G. (2009) Data assimilation: the ensemble Kalman filter
Springer-Verlag.

Anderson, J. L. (2001) An Ensemble Adjustment Kalman Filter for Data
Assimilation Monthly Weather Review 129:2884--2903
}
\seealso{
Other \pkg{spatpomp} parameter estimation methods: \code{\link{hippie}}
}
\author{
Kidus Asfaw
}
\concept{\pkg{spatpomp} parameter estimation methods}
\concept{particle filtering methods}
