% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senkf2.R
\name{senkf2}
\alias{senkf2}
\alias{senkf2,ANY-method}
\alias{senkf2,missing-method}
\title{Ensemble Kalman filters}
\arguments{
\item{Np}{the number of particles to use.}
}
\value{
An object of class \sQuote{senkfd_spatPomp}.
}
\description{
A function to perform filtering using the ensemble Kalman filter of Evensen, G. (1994)
}
\references{
Evensen, G. (1994) Sequential data assimilation with a
nonlinear quasi-geostrophic model using Monte Carlo methods to forecast
error statistics Journal of Geophysical Research: Oceans 99:10143--10162

Evensen, G. (2009) Data assimilation: the ensemble Kalman filter
Springer-Verlag.

Anderson, J. L. (2001) An Ensemble Adjustment Kalman Filter for Data
Assimilation Monthly Weather Review 129:2884--2903
}
\seealso{
Other particle filtering methods: \code{\link{senkf}}

Other \pkg{spatPomp} parameter estimation methods: \code{\link{hippie}},
  \code{\link{senkf}}
}
\concept{\pkg{spatPomp} parameter estimation methods}
\concept{particle filtering methods}
