% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show-spatPomp}
\alias{show-spatPomp}
\alias{show,spatPomp-method}
\title{Show methods}
\usage{
\S4method{show}{spatPomp}(object)
}
\arguments{
\item{object}{a \code{spatPomp} object}
}
\description{
Display the object, according to its class.
}
