% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undefined.R
\name{undefined}
\alias{undefined}
\alias{undefined,NULL-method}
\alias{undefined,ANY-method}
\alias{undefined,missing-method}
\alias{undefined,pomp_fun-method}
\alias{undefined,rprocPlugin-method}
\title{Undefined}
\usage{
undefined(object, ...)

\S4method{undefined}{NULL}(object, ...)

\S4method{undefined}{ANY}(object, ...)

\S4method{undefined}{missing}(object, ...)

\S4method{undefined}{pomp_fun}(object, ...)

\S4method{undefined}{rprocPlugin}(object, ...)
}
\arguments{
\item{object}{object to test.}

\item{...}{currently ignored.}
}
\value{
Returns \code{TRUE} if the \pkg{pomp} workhorse method is undefined,
\code{FALSE} if it is defined,
and \code{NA} if the question is inapplicable.
}
\description{
Check for undefined methods.
}
\keyword{internal}
