% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_names.R
\name{unit_names}
\alias{unit_names}
\alias{unit_names-spatPomp}
\alias{unit_names,spatPomp-method}
\title{Unit names of a spatiotemporal model}
\usage{
\S4method{unit_names}{spatPomp}(x)
}
\arguments{
\item{x}{a \code{spatPomp} object}
}
\description{
\code{unit_names} outputs the contents of the \code{unit_names} slot
of a \code{spatPomp} object. The order in which the spatial units
appear in the output vector determines the order in which latent
states and observations for the spatial units are stored.
}
