% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_dmeasure.R
\name{vec_dmeasure.internal}
\alias{vec_dmeasure.internal}
\title{Evaluate the unit measurement model density function for each unit}
\usage{
vec_dmeasure.internal(object, y, x, times, params, log = FALSE,
  .gnsi = TRUE, ...)
}
\description{
Evaluate the unit measurement model density function for each unit
}
