% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_rmeasure.R
\name{vec_rmeasure}
\alias{vec_rmeasure}
\alias{vec_rmeasure-spatPomp}
\alias{vec_rmeasure,spatPomp-method}
\title{Vector of simulated measurements}
\usage{
\S4method{vec_rmeasure}{spatPomp}(object, x, times, params, ...)
}
\arguments{
\item{object}{a \code{spatPomp} object}

\item{x}{numeric; state at which measurements are simulated}

\item{times}{numeric; time at which measurements are simulated}

\item{params}{numeric; parameter set at which measurements are simulated}

\item{\dots}{additional parameters will be ignored}
}
\description{
Simulate from the unit measurement model density function for each unit
}
