% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_spatPomp.R
\name{as_spatPomp}
\alias{as_spatPomp}
\alias{coerce-pomp-spatPomp}
\alias{coerce,pomp,spatPomp-method}
\title{Coerce to spatPomp}
\description{
Convert to class spatPomp object
}
\details{
When \code{object} is a simple \sQuote{pomp} object,
construct and return a one-dimensional \sQuote{spatPomp} object.
}
