% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asif.R
\docType{methods}
\name{asif}
\alias{asif}
\alias{asif-spatpomp}
\alias{asif,spatpomp-method}
\title{Adapted Simulation Island Filter (ASIF)}
\usage{
\S4method{asif}{spatpomp}(object, params, Np, nbhd, islands,
  tol = (1e-18)^9, max.fail = Inf, pred.mean = FALSE,
  pred.var = FALSE, filter.mean = FALSE, filter.traj = FALSE,
  save.states = FALSE, save.params = FALSE,
  verbose = getOption("verbose"), ...)
}
\arguments{
\item{object}{A \code{spatpomp} object.}

\item{params}{A parameter set for the spatiotemporal POMP.}

\item{Np}{The number of particles used within each island for the adapted simulations.}

\item{nbhd}{A function which when given a point in space-time \code{(d,n)} will output a 0-1 matrix of points in space-time considered to be
in the neighborhood of \code{(d,n)}.}

\item{islands}{The number of islands for the adapted simulations.}

\item{tol}{
    positive numeric scalar; particles with likelihood less than \code{tol} are considered to be incompatible with the data.
    See the section on \emph{Filtering Failures} below for more information.
  }

\item{max.fail}{
    integer; the maximum number of filtering failures allowed (see below).
    If the number of filtering failures exceeds this number, execution will terminate with an error.
    By default, \code{max.fail} is set to infinity, so no error can be triggered.
  }

\item{pred.mean}{
    logical; if \code{TRUE}, the prediction means are calculated for the state variables and parameters.
  }

\item{pred.var}{
    logical; if \code{TRUE}, the prediction variances are calculated for the state variables and parameters.
  }

\item{filter.mean}{
    logical; if \code{TRUE}, the filtering means are calculated for the state variables and parameters.
  }

\item{filter.traj}{
    logical; if \code{TRUE}, a filtered trajectory is returned for the state variables and parameters.
  }

\item{save.states}{
    logical.
    If \code{save.states=TRUE}, the state-vector for each particle at each time is saved in the \code{saved.states} slot of the returned \code{\link[=pfilterd.pomp-class]{pfilterd.pomp}} object.
    If \code{save.params=TRUE}, the parameter-vector for each particle at each time is saved in the \code{saved.params} slot of the returned \code{\link[=pfilterd.pomp-class]{pfilterd.pomp}} object.
  }

\item{save.params}{
    logical.
    If \code{save.states=TRUE}, the state-vector for each particle at each time is saved in the \code{saved.states} slot of the returned \code{\link[=pfilterd.pomp-class]{pfilterd.pomp}} object.
    If \code{save.params=TRUE}, the parameter-vector for each particle at each time is saved in the \code{saved.params} slot of the returned \code{\link[=pfilterd.pomp-class]{pfilterd.pomp}} object.
  }

\item{verbose}{
    logical; if \code{TRUE}, progress information is reported as \code{pfilter} works.
  }

\item{...}{
    Any additional arguments given to \code{pomp} will be made available to each of the basic components.
    To prevent errors due to misspellings, a warning is issued if any such arguments are detected.
  }
}
\value{
Upon successful completion, \code{asif} returns an object of class
\sQuote{asifd.pomp}.
}
\description{
An algorithm for estimating the filter distribution of a spatiotemporal partially-observed Markov process (SpatPOMP for short).
Running \code{asif} causes the algorithm to run independent island jobs which each yield an imperfect adapted simulation. Simulating from the "adapted filter"
distribution runs into a curse of dimensionality (COD) problem, which is mitigated by keeping particles in each island close to each other through resampling down
to one particle per island at each observation time point.
The adapted simulations are then weighted in a way that tries to avert COD by making a weak coupling assumption to get an approximate filter distribution.
As a by-product, we also get a biased estimate of the likelihood of the data.
}
\section{Methods}{

The following methods are available for such an object:
\describe{
\item{\code{\link{loglik}}}{ yields a biased estimate of the log-likelihood of the data under the model. }
}
}

\seealso{
Other particle filter methods: \code{\link{girf}},
  \code{\link{hippie}}, \code{\link{igirf}}

Other \pkg{spatpomp} filtering methods: \code{\link{girf}},
  \code{\link{igirf}}
}
\concept{\pkg{spatpomp} filtering methods}
\concept{particle filter methods}
