% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enkf.R
\name{enkf}
\alias{enkf}
\alias{enkf,ANY-method}
\alias{enkf,missing-method}
\alias{enkf-spatPomp}
\alias{enkf,spatPomp-method}
\title{Generalized Ensemble Kalman filter (EnKF)}
\usage{
\S4method{enkf}{spatPomp}(data, Np, ..., verbose = getOption("verbose", FALSE))
}
\arguments{
\item{data}{A \code{spatPomp} object.}

\item{Np}{The number of Monte Carlo particles used to approximate the filter distribution.}
}
\value{
An object of class \sQuote{enkfd_spatPomp}.
}
\description{
A function to perform filtering using the ensemble Kalman filter of Evensen, G. (1994).
This function is generalized to allow for an measurement covariance matrix that varies over time.
This is useful if the measurement model varies with the state.
}
\examples{
# Create a simulation of a Brownian motion
b <- bm(U=6, N=10)

# Run EnKF
enkfd_bm <- enkf(b, Np = 100)

# Get a likelihood estimate
logLik(enkfd_bm)
}
\references{
\Evensen1994

\Evensen2009

\Anderson2001
}
\seealso{
Other particle filter methods: 
\code{\link{abfir}()},
\code{\link{abf}()},
\code{\link{bpfilter}()},
\code{\link{girf}()},
\code{\link{ienkf}()},
\code{\link{igirf}()}

Other \pkg{spatPomp} filtering methods: 
\code{\link{abfir}()},
\code{\link{abf}()},
\code{\link{bpfilter}()},
\code{\link{girf}()},
\code{\link{igirf}()}
}
\concept{\pkg{spatPomp} filtering methods}
\concept{particle filter methods}
