% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enkf.R
\name{enkf}
\alias{enkf}
\alias{enkf,ANY-method}
\alias{enkf,missing-method}
\alias{enkf-spatPomp}
\alias{enkf,spatPomp-method}
\title{Generalized Ensemble Kalman filter}
\usage{
\S4method{enkf}{spatPomp}(data, Np, ..., verbose = getOption("verbose", FALSE))
}
\arguments{
\item{data}{either a data frame holding the spatiotemporal data,
or an object of class \sQuote{spatPomp},
i.e., the output of another \pkg{spatPomp} calculation.}

\item{Np}{the number of particles to use.}

\item{...}{additional arguments supply new or modify existing model characteristics or components.
See \code{\link[pomp]{pomp}} for a full list of recognized arguments.

When named arguments not recognized by \code{\link[pomp]{pomp}} are provided, these are made available to all basic components via the so-called \dfn{userdata} facility.
This allows the user to pass information to the basic components outside of the usual routes of covariates (\code{covar}) and model parameters (\code{params}).
See \link[pomp:userdata]{?userdata} for information on how to use this facility.}
}
\value{
An object of class \sQuote{enkfd_spatPomp}.
}
\description{
A function to perform filtering using the ensemble Kalman filter of Evensen, G. (1994).
This function is generalized to allow for an R matrix that varies over time.
This is useful if the measurement model varies with the state.
}
\references{
Evensen, G. (1994) Sequential data assimilation with a
nonlinear quasi-geostrophic model using Monte Carlo methods to forecast
error statistics Journal of Geophysical Research: Oceans 99:10143--10162

Evensen, G. (2009) Data assimilation: the ensemble Kalman filter
Springer-Verlag.

Anderson, J. L. (2001) An Ensemble Adjustment Kalman Filter for Data
Assimilation Monthly Weather Review 129:2884--2903
}
\seealso{
Other \pkg{spatPomp} parameter estimation methods: 
\code{\link{ienkf-spatPomp}}
}
\concept{\pkg{spatPomp} parameter estimation methods}
\concept{particle filtering methods}
