% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eunit_measure.R
\name{eunit_measure}
\alias{eunit_measure}
\title{eunit_measure}
\usage{
\S4method{eunit_measure}{spatPomp}(object, x, unit, time, params, Np = 1, log = FALSE, gnsi = TRUE)
}
\arguments{
\item{object}{An object of class \code{spatPomp}}

\item{x}{A state vector for all units}

\item{unit}{The unit for which to evaluate the expectation}

\item{time}{The time for which to evaluate the expectation}

\item{params}{parameters at which to evaluate the unit expectation}
}
\value{
A matrix with the unit expectation
}
\description{
\code{eunit_measure} evaluates the expectation of a unit's observation given the entire state
}
\examples{
b <- bm(U=3)
s <- states(b)[,1,drop=FALSE]
rownames(s) -> rn
dim(s) <- c(3,1,1)
dimnames(s) <- list(variable=rn, rep=NULL)
p <- coef(b); names(p) -> rnp
dim(p) <- c(length(p),1); dimnames(p) <- list(param=rnp)
o <- obs(b)[,1,drop=FALSE]
eunit_measure(b, x=s, unit=2, time=1, params=p)
}
