% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girf.R
\docType{methods}
\name{girf}
\alias{girf}
\alias{girf-spatPomp}
\alias{girf,spatPomp-method}
\alias{girf-girfd_spatPomp}
\alias{girf,girfd_spatPomp-method}
\title{Guided intermediate resampling filter (GIRF)}
\usage{
\S4method{girf}{spatPomp}(data, Np, Ninter, lookahead, Nguide, h,
  theta.to.v, v.to.theta, tol = 1e-17, max.fail = Inf,
  save.states = FALSE, ..., verbose = getOption("verbose", FALSE))

\S4method{girf}{girfd_spatPomp}(data, Np, tol, Ninter, Nguide, lookahead,
  h, theta.to.v, v.to.theta, ..., verbose = getOption("verbose", FALSE))
}
\arguments{
\item{data}{either a data frame holding the spatiotemporal data,
or an object of class \sQuote{spatPomp},
i.e., the output of another \pkg{spatPomp} calculation.}

\item{Np}{The number of particles used.}

\item{Ninter}{the number of intermediate resampling timepoints.}

\item{lookahead}{The number of future observations included in the guide function.}

\item{Nguide}{The number of simulations used to estimate state process uncertainty for each particle.}

\item{h}{A function with template \code{function(state.vec, param.vec)} which returns an approximation to the expected observed value given a latent state and parameter vector.}

\item{theta.to.v}{A function with template \code{function(state.vec, param.vec)} which returns an approximation to the variance of an observation given a latent state and parameter vector.}

\item{v.to.theta}{A function with template \code{function(var, state.vec, param.vec)} which returns a parameter vector having observation noise consistent with variance \code{var} at latent state \code{state.vec} with other paramters given by \code{param.vec}.}

\item{...}{additional arguments supply new or modify existing model characteristics or components.
See \code{\link{pomp}} for a full list of recognized arguments.

When named arguments not recognized by \code{\link{pomp}} are provided, these are made available to all basic components via the so-called \dfn{userdata} facility.
This allows the user to pass information to the basic components outside of the usual routes of covariates (\code{covar}) and model parameters (\code{params}).
See \link[=userdata]{?userdata} for information on how to use this facility.}

\item{verbose}{logical; if \code{TRUE}, diagnostic messages will be printed to the console.}

\item{object}{A \code{spatPomp} object.}

\item{params}{A parameter set for the spatiotemporal POMP.}
}
\value{
Upon successful completion, \code{girf} returns an object of class
\sQuote{girfd.spatPomp}.
}
\description{
An implementation of the algorithm of Park and Ionides (2019),
following the pseudocode in Asfaw, Ionides and King (2019).
}
\section{Methods}{

The following methods are available for such an object:
\describe{
\item{\code{\link{logLik}}}{ yields a biased estimate of the log-likelihood of the data under the model. }
}
}

\references{
\park2019

\asfaw2019
}
\seealso{
Other particle filter methods: \code{\link{asifir}},
  \code{\link{asif}}, \code{\link{hippie}},
  \code{\link{igirf}}

Other \pkg{spatPomp} filtering methods: \code{\link{asifir}},
  \code{\link{asif}}, \code{\link{igirf}}
}
\concept{\pkg{spatPomp} filtering methods}
\concept{particle filter methods}
