% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatPomp_methods.R
\name{logLik-abfd_spatPomp}
\alias{logLik-abfd_spatPomp}
\alias{logLik,abfd_spatPomp-method}
\alias{logLik-iabfd_spatPomp}
\alias{logLik,iabfd_spatPomp-method}
\alias{logLik-abfird_spatPomp}
\alias{logLik,abfird_spatPomp-method}
\alias{logLik-igirfd_spatPomp}
\alias{logLik,igirfd_spatPomp-method}
\title{loglik}
\usage{
\S4method{logLik}{abfd_spatPomp}(object)

\S4method{logLik}{iabfd_spatPomp}(object)

\S4method{logLik}{abfird_spatPomp}(object)

\S4method{logLik}{igirfd_spatPomp}(object)
}
\description{
loglik

loglik

loglik

loglik
}
