% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatPomp_methods.R
\docType{methods}
\name{logLik-girfd_spatPomp}
\alias{logLik-girfd_spatPomp}
\alias{logLik,girfd_spatPomp-method}
\alias{logLik-asifd.spatPomp}
\alias{logLik,asifd.spatPomp-method}
\alias{logLik-asifird_spatPomp}
\alias{logLik,asifird_spatPomp-method}
\alias{logLik-igirfd_spatPomp}
\alias{logLik,igirfd_spatPomp-method}
\title{loglik}
\usage{
\S4method{logLik}{girfd_spatPomp}(object)

\S4method{logLik}{asifd_spatPomp}(object)

\S4method{logLik}{asifird_spatPomp}(object)

\S4method{logLik}{igirfd_spatPomp}(object)
}
\description{
loglik

loglik

loglik

loglik
}
