% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{logLik}
\alias{logLik}
\alias{logLik-girfd_spatPomp}
\alias{logLik,girfd_spatPomp-method}
\alias{logLik-bpfilterd_spatPomp}
\alias{logLik,bpfilterd_spatPomp-method}
\alias{logLik-abfd_spatPomp}
\alias{logLik,abfd_spatPomp-method}
\alias{logLik-iubfd_spatPomp}
\alias{logLik,iubfd_spatPomp-method}
\alias{logLik-abfird_spatPomp}
\alias{logLik,abfird_spatPomp-method}
\alias{logLik-igirfd_spatPomp}
\alias{logLik,igirfd_spatPomp-method}
\title{Log likelihood extractor}
\usage{
\S4method{logLik}{girfd_spatPomp}(object)

\S4method{logLik}{bpfilterd_spatPomp}(object)

\S4method{logLik}{abfd_spatPomp}(object)

\S4method{logLik}{iubfd_spatPomp}(object)

\S4method{logLik}{abfird_spatPomp}(object)

\S4method{logLik}{igirfd_spatPomp}(object)
}
\arguments{
\item{object}{fitted model object}
}
\value{
a numeric which is the estimated log likelihood
}
\description{
Extracts the estimated log likelihood from a fitted model
}
