% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz.R
\name{lorenz}
\alias{lorenz}
\title{Lorenz 96 spatPomp generator}
\usage{
lorenz(
  U = 5,
  N = 100,
  dt = 0.01,
  dt_obs = 0.5,
  regular_params = c(F = 8, sigma = 1, tau = 1)
)
}
\arguments{
\item{U}{A length-one numeric signifying the number of spatial units for the process.}

\item{N}{A length-one numeric signifying the number of observations.}

\item{dt}{A length-one numeric giving the Euler time step for the numerical solution.}

\item{dt_obs}{A length-one numeric giving the time between observations.}
}
\value{
A spatPomp object with the specified dimension and time steps.
}
\description{
Generate a spatPomp object representing a \code{U}-dimensional stochastic Lorenz 96 process with
\code{N} measurements made at times \eqn{t_n = n * dt_{obs}}, simulated using an Euler method
with time increment \code{dt}.
}
\examples{
l <- lorenz(U=5, N=100, dt=0.01, dt_obs=1)
# See all the model specifications of the object
spy(l)
}
