% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles.R
\name{measles}
\alias{measles}
\title{Measles in UK spatPomp generator}
\usage{
measles(
  U = 6,
  dt = 2/365,
  fixed_ivps = TRUE,
  shared_ivps = TRUE,
  S_0 = 0.032,
  E_0 = 5e-05,
  I_0 = 4e-05
)
}
\arguments{
\item{U}{A length-one numeric signifying the number of cities to be represented in the spatPomp object.}
}
\value{
A spatPomp object.
}
\description{
Generate a spatPomp object for measles in the top-\code{U} most populous cities in England.
The model is adapted from He et al. (2010) with gravity transport following Park and Ionides (2019).
The data in the object is simulated using the process and measurement models of He et al. (2010).
}
\examples{
m <- measles(U = 7)
# See all the model specifications of the object
spy(m)
}
