% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles.R
\name{measles}
\alias{measles}
\title{Measles in UK spatpomp generator}
\usage{
measles(D = 6)
}
\arguments{
\item{D}{A length-one numeric signifying the number of cities to be represented in the spatpomp object.}
}
\value{
A spatpomp object.
}
\description{
Generate a spatpomp object for measles in the top-\code{D} most populous cities in England.
}
\examples{
measles(7)
}
