% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ou.R
\name{ou}
\alias{ou}
\title{Linear-Gaussian spatPomp generator}
\usage{
ou(U = 5, N = 100)
}
\arguments{
\item{U}{A length-one numeric signifying dimension of the process.}

\item{N}{A length-one numeric signifying the number of time steps to evolve the process.}
}
\value{
A spatPomp object with the specified dimension and time steps.
}
\description{
Generate a spatPomp object representing a \code{U}-dimensional discrete-time
Ornstein-Uhlenbeck process with \code{N} time steps.
}
\examples{
ou(5, 100)

}
