% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print-spatPomp}
\alias{print,spatPomp-method}
\title{Print methods}
\usage{
\S4method{print}{spatPomp}(x)
}
\arguments{
\item{x}{a \code{spatPomp} object}
}
\description{
Prints its argument.
}
