% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatPomp_class.R
\docType{class}
\name{spatPomp-class}
\alias{spatPomp-class}
\title{An S4 class to represent a spatiotemporal POMP model and data.}
\description{
An S4 class to represent a spatiotemporal POMP model and data.
}
\section{Slots}{

\describe{
\item{\code{unit_names}}{A vector containing the spatial units of a spatiotemporal POMP.}

\item{\code{unit_statenames}}{A vector containing the state names such that appending the unit indices to the
unit statenames will result in the each unit's corresponding states.}

\item{\code{unit_obsnames}}{A vector of observation types for a spatial unit.}

\item{\code{eunit_measure}}{A pomp_fun representing the expected measurement for each spatial unit given its states.}

\item{\code{dunit_measure}}{A pomp_fun representing the unit measurement density for each spatial unit.}

\item{\code{runit_measure}}{A pomp_fun representing the unit observation simulator.}
}}

