% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatPomp_class.R
\docType{class}
\name{spatPomp-class}
\alias{spatPomp-class}
\title{An S4 class to represent a spatiotemporal POMP model and data.}
\description{
An S4 class to represent a spatiotemporal POMP model and data.
}
\section{Slots}{

\describe{
\item{\code{units}}{A vector containing the spatial units of a spatiotemporal POMP.}

\item{\code{unit_index}}{A vector with the spatial units as values and the indices for the units as the names.}

\item{\code{unit_statenames}}{A vector containing the state names such that appending the unit indices to the
unit statenames will result in the each unit's corresponding states.}

\item{\code{global_statenames}}{A vector containing the statenames shared by all spatial units.}

\item{\code{obstypes}}{A vector of observation types for a spatial unit.}

\item{\code{emeasure}}{A pomp_fun representing the expected measurement for each spatial unit given its states.}

\item{\code{unit_dmeasure}}{A pomp_fun representing the unit measurement density for each spatial unit.}

\item{\code{unit_rmeasure}}{A pomp_fun representing the unit observation simulator.}
}}

