% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatPomp_methods.R
\docType{methods}
\name{spatPomp_Csnippet}
\alias{spatPomp_Csnippet}
\title{spatPomp_Csnippet}
\usage{
\S4method{spatPomp_Csnippet}{character}(object, unit_statenames,
  unit_covarnames)
}
\description{
spatPomp_Csnippet
}
