% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatPomp_methods.R
\name{unit_names}
\alias{unit_names}
\title{Unit names of a spatiotemporal model}
\usage{
\S4method{unit_names}{spatPomp}(x, ...)
}
\arguments{
\item{...}{additional arguments supply new or modify existing model characteristics or components.
See \code{\link[pomp]{pomp}} for a full list of recognized arguments.

When named arguments not recognized by \code{\link[pomp]{pomp}} are provided, these are made available to all basic components via the so-called \dfn{userdata} facility.
This allows the user to pass information to the basic components outside of the usual routes of covariates (\code{covar}) and model parameters (\code{params}).
See \link[pomp:userdata]{?userdata} for information on how to use this facility.}
}
\description{
\code{unit_names} outputs the contents of the \code{unit_names} slot
of a \code{spatPomp} object. The order in which the spatial units
appear in the output vector determines the order in which latent
states and observations for the spatial units are stored.
}
